+++ backend/app/src/main/java/com/ecomerce/spring/app/model/Usuario.java
@@ -14,7 +14,7 @@ import jakarta.persistence.GenerationType;
 public class Usuario {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_usuario;
+    private Long id_usuario;
 
     private String nombre_completo;
     private String email;
@@ -26,10 +26,10 @@ public class Usuario {
     private LocalDateTime fechaRegistro;
 
     // Getters y setters
-    public int getId_usuario() {
+    public Long getId_usuario() {
         return id_usuario;
     }
-    public void setId_usuario(int id_usuario) {
+    public void setId_usuario(Long id_usuario) {
         this.id_usuario = id_usuario;
     }
